Make spawners a module
